Moved GET /current to the UserController as it's the only method without authorization.
